# Eagerly require `date`