--- a/test/psych/test_date_time.rb
+++ b/test/psych/test_date_time.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 require_relative 'helper'
-require 'date'
 
 module Psych
   class TestDateTime < TestCase