@byroot 2024-11-19
You can probably use a subprocess.