@tdeo 2024-11-19
> You can probably use a subprocess.

I was able indeed to add a test case with this technique, thanks for the suggestion