@byroot 2024-11-19
Right, sorry Aaron, I should have explained.

In `YAMLTree#visit_String`, there's this condition to figure if a string should be quoted:

```ruby
        elsif not String === @ss.tokenize(o) or /\A0[0-7]*[89]/.match?(o)
          style = Nodes::Scalar::SINGLE_QUOTED
```

So we end up in:

```ruby
class_loader.date.strptime(string, '%F', Date::GREGORIAN)
```

Which further end up trying `ClassLoader#resolve` of "Date", and this uses `path2class`, so it's resolved as essentially `::Date`, so the scoped `autoload` I introduced isn't triggered in this case.

Hence why the initial version of this PR was to move the autoload at the top level, but at that point I think we should just eager load.