@tenderlove 2024-11-20
@byroot ah, gotcha.  I somewhat worry this might have downstream effects (maybe on Bundler or RubyGems).  But I like the change so I'll merge it.